+++ Frontend/Vue/src/router/index.js
@@ -1,6 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import BaseLayout from '@/layouts/BaseLayout.vue';
 import Dashboard from '@/views/Dashboard.vue';
+import CreateWorkout from '@/views/CreateWorkout.vue';
 
 const routes = [
   {
@@ -10,6 +11,10 @@ const routes = [
       {
         path: '',
         component: Dashboard
+      }, 
+      {
+        path: 'create-workout',
+        component: CreateWorkout
       }
     ]
   }