+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## Jul 28, 2024
+- Updated Frontend: Added `Create Workout Routine` Component.
 
 ## Jul 26, 2024
 - Commit first stable project files.